2001-02-14  Carlos Jimenez  <jimenez@rss.chalmers.se>
	* arts-0-0-123

	* maths_func.h, maths_func.cc: Added the function *chol* to do a
	Choleski factorization

	* demonstrate_vecmat.cc: Included a test case running *chol*


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@788 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/arts.h
@@ -88,7 +88,7 @@
 /** The type to use for all floating point numbers. You should never
     use float or double explicitly, unless you have a very good
     reason. Always use this type instead.  */
-typedef float Numeric;
+typedef double Numeric;
 
 //--------------------< Set integer type >--------------------
 /** The type to use for all integer numbers and indices. You should never
